requirements.txt
# Core
Django==2.0.3
djangorestframework==3.7.7
django_extensions==2.0.3
python-dateutil====2.7.0

# Tests
pytest==3.4.2
pytest-django==3.1.2
model_mommy==1.5.1
faker==0.8.11
